day5/cloud/code/pyproject.toml
[project]
name = "psu_de"
version = "0.1.0"
description = ""
package-mode = false
authors = [
    {name = "Thanathip Limna",email = "thanathip.l@coe.psu.ac.th"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "beanie (>=1.30.0,<2.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "dash (>=3.0.4,<4.0.0)",
    "pandas (>=2.3.0,<3.0.0)",
    "numpy (>=2.3.0,<3.0.0)",
    "plotly[express] (>=6.1.2,<7.0.0)",
    "python-dateutil (>=2.9.0.post0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
